move HGT class into its own source file
